change get group enrollment configuration web service
